First commit for Adversarial World!
This project is the implementation effort for an extension of the
"Wumpus World"-style game, often used for academia in AI. The goal of
this project was to create two agents in a modified version of the
Wumpus World game in which two agents traverse the game map with their
own unique sets of rules and environment interactions. This project was
intended to test how well the agents learn from the environment and each
other as the project was set to have a series of "matches" played out by
the two agents.